Restructures and makes ui folder
