+++ src/components/Cart/CartItem/CartItem.jsx
@@ -1,6 +1,6 @@
 import useCart from "../../../hooks/useCart";
 import { Link } from "react-router-dom";
-import Button from "../../Button";
+import Button from "../../ui/Button";
 import {
   ProductContainer,
   ImageContainer,
