+++ src/components/Checkout/Checkout.jsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import useCart from "../../hooks/useCart";
-import Breadcrumbs from "../Breadcrumbs";
+import Breadcrumbs from "../ui/Breadcrumbs";
 import OrderSummary from "./CheckoutForm/OrderSummary";
 import CheckoutForm from "./CheckoutForm";
 import { CheckoutContainer, Container, Heading } from "./Checkout.styled";
