+++ src/components/Checkout/CheckoutForm/CheckoutForm.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import PaymentOptions from "./PaymentOptions/PaymentOptions";
-import Button from "../../Button";
+import Button from "../../ui/Button";
 import { getTotalPrice } from "../../../utils";
 
 import { StyledForm } from "./CheckoutForm.styled";
