+++ src/components/Product/index.jsx
@@ -1,0 +1,1 @@
+export { default } from "./Product";
