--- src/components/SingleProduct/index.jsx
@@ -1,1 +1,0 @@
-export { default } from "./SingleProduct";
